# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: Opis backendu
  version: 0.0.1
servers:
- url: http://localhost:8080
paths:
  /api/v1/new:
    post:
      summary: Tworzenie nowego stashu
      description: Endpoint do tworzenia nowego kodu (i nowego stashu).
      responses:
        '201':
          $ref: '#/components/responses/NewStashCreated'
        '507':
          $ref: '#/components/responses/InsufficientStorage'
  /api/v1/view/{code}:
    get:
      summary: Podgląd zawartości stashu
      description: Endpoint do wyświetlania zawartości stashu dostępnego pod danym kodem.
      parameters:
      - in: path
        name: code
        required: true
        schema:
          type: string
        description: Kod stashu
      responses:
        '200':
          $ref: '#/components/responses/StashViewSuccess'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/prolong/{code}:
    post:
      summary: Przedłużenie ważności stashu
      description: Endpoint do przedłużania ważności stashu dostępnego pod danym kodem.
        Po przedłużeniu ważności stashu, data i czas, do którego stash jest ważny,
        zostanie zaktualizowana do nowej wartości podanej w body żądania.
      parameters:
      - in: path
        name: code
        required: true
        schema:
          type: string
        description: Kod stashu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                new_expiration_date:
                  type: string
                  format: date-time
                  description: Nowa data i czas, do którego stash jest ważny
                  example: '2024-12-31T23:59:59Z'
      responses:
        '204':
          $ref: '#/components/responses/SuccessNoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/make-readonly/{code}:
    post:
      summary: Ustawienie stashu jako read-only
      description: Endpoint do ustawiania stashu jako tylko do odczytu. Po ustawieniu
        stashu jako tylko do odczytu, nie będzie można już uploadować nowych plików
        ani folderów do tego stashu, ale nadal będzie można przeglądać jego zawartość.
      parameters:
      - in: path
        name: code
        required: true
        schema:
          type: string
        description: Kod stashu
      responses:
        '204':
          $ref: '#/components/responses/SuccessNoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/upload/{code}:
    post:
      summary: Upload pliku lub folderu
      description: Endpoint do uploadowania pliku lub folderu do stashu dostępnego pod
        danym kodem. Foldery powinny być spakowane do formatu zip przed uploadem,
        a następnie będą rozpakowywane po stronie backendu. Jeżeli stash jest tylko
        do odczytu, upload nie będzie możliwy.
      parameters:
      - in: path
        name: code
        required: true
        schema:
          type: string
        description: Kod stashu
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                is_directory:
                  type: boolean
                  description: Informacja, czy uploadowany element jest katalogiem
                    (true) czy plikiem (false). Jeżeli jest to katalog, powinien być
                    spakowany do formatu zip. Jeśli jest to plik, powinien być uploadowany
                    bezpośrednio.
                parent_path:
                  type: string
                  description: Ścieżka do katalogu w staszu, do którego ma być uploadowany
                    plik (np. "/folder1/folder2"). Może być puste dla katalogu głównego.
                    Nie może zawierać ../ lub //
                  pattern: ^(/[a-zA-Z0-9._\- ]+)+/?$
                  minLength: 0
                  maxLength: 1024
                name:
                  type: string
                  description: Nazwa pliku lub folderu (np. "file.txt" lub "folder").
                    Powinien zawierać rozszerzenie, jeśli jest to plik (np. "file.txt").
                    Jeśli jest to folder, nie powinien zawierać rozszerzenia (np.
                    "folder").
                data:
                  type: string
                  format: binary
                  description: Plik lub folder do uploadowania (jeśli jest to folder,
                    powinien być spakowany do formatu zip)
      responses:
        '204':
          $ref: '#/components/responses/SuccessNoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '507':
          $ref: '#/components/responses/InsufficientStorage'
  /api/v1/download/{code}:
    get:
      summary: Pobieranie zasobu ze stashu
      description: Endpoint do pobierania pliku lub folderu ze stashu dostępnego pod danym
        kodem. Domyślnie plik jest zwracany w oryginalnej postaci; jeśli pobierany
        element jest katalogiem, zostanie spakowany do formatu zip przed wysłaniem.
        Jeśli w zapytaniu podano `format=zip`, zasób (plik lub katalog) zostanie zwrócony
        jako archiwum zip.
      parameters:
      - in: path
        name: code
        required: true
        schema:
          type: string
        description: Kod stashu
      - in: query
        name: path
        required: true
        schema: &id001
          type: string
          pattern: ^(/[a-zA-Z0-9._\- ]+)+/?$
          minLength: 1
          maxLength: 1024
          description: Ścieżka do pliku lub folderu w staszu (np. "/folder1/folder2/file.txt").
            Ścieżka musi zaczynać się od "/" i zawierać tylko znaki alfanumeryczne,
            podkreślenia, myślniki i spacje. Ścieżki zawierające ../ lub // są niedozwolone.
        description: Ścieżka do pliku lub folderu w staszu, który ma zostać pobrany
          (np. "/folder1/folder2/file.txt")
      - in: query
        name: format
        required: false
        schema:
          type: string
          enum:
          - zip
        description: Opcjonalny format odpowiedzi. Ustaw `zip`, aby otrzymać zasób
          spakowany jako plik .zip.
      responses:
        '200':
          $ref: '#/components/responses/FileDownloadSuccess'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/delete/{code}:
    delete:
      summary: Usuwanie pliku lub folderu
      description: Endpoint do usuwania pliku lub folderu ze stashu dostępnego pod danym
        kodem. Jeżeli usuwany element jest katalogiem, zostanie usunięty wraz z całą
        swoją zawartością. Jeżeli stash jest tylko do odczytu, nie będzie można usuwać
        jego zawartości.
      parameters:
      - in: path
        name: code
        required: true
        schema:
          type: string
        description: Kod stashu
      - in: query
        name: path
        required: true
        schema: *id001
        description: Ścieżka do pliku lub folderu w staszu, który ma zostać usunięty
          (np. "/folder1/folder2/file.txt")
      responses:
        '204':
          $ref: '#/components/responses/SuccessNoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/rename/{code}:
    post:
      summary: Zmiana nazwy zasobu
      description: Endpoint do zmiany nazwy pliku lub folderu w stashu dostępnym pod danym
        kodem. Jeżeli stash jest tylko do odczytu, nie będzie można zmieniać nazw
        jego zawartości.
      parameters:
      - in: path
        name: code
        required: true
        schema:
          type: string
        description: Kod stashu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                old_path:
                  type: string
                  description: Obecna ścieżka do pliku lub folderu (np. "/folder1/folder2/file.txt").
                    Ścieżka musi zaczynać się od "/" i nie może zawierać ../ lub //
                  pattern: ^(/[a-zA-Z0-9._\- ]+)+/?$
                  minLength: 1
                  maxLength: 1024
                new_name:
                  type: string
                  description: Nowa nazwa (wraz z rozszerzeniem, jeśli jest to plik)
                    pliku (np. "new_file.txt" lub "new_folder"). Jeśli jest to folder,
                    nie powinna zawierać rozszerzenia (np. "new_folder").
      responses:
        '204':
          $ref: '#/components/responses/SuccessNoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/new-folder/{code}:
    post:
      summary: Tworzenie nowego folderu
      description: Endpoint do tworzenia nowego, pustego folderu w stashu dostępnym pod
        danym kodem. Jeżeli stash jest tylko do odczytu, nie będzie można tworzyć
        nowych folderów.
      parameters:
      - in: path
        name: code
        required: true
        schema:
          type: string
        description: Kod stashu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                parent_path:
                  type: string
                  description: Ścieżka do katalogu, w którym ma być utworzony nowy
                    folder (np. "/folder1"). Może być puste dla katalogu głównego.
                    Nie może zawierać ../ lub //
                  pattern: ^(/[a-zA-Z0-9._\- ]+)+/?$
                  minLength: 0
                  maxLength: 1024
                name:
                  type: string
                  description: Nazwa nowego folderu (np. "new_folder").
      responses:
        '204':
          $ref: '#/components/responses/SuccessNoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '507':
          $ref: '#/components/responses/InsufficientStorage'
  /api/v1/update-category/{code}:
    post:
      summary: Aktualizacja kategorii pliku
      description: Endpoint do aktualizowania kategorii pliku w stashu dostępnym pod danym
        kodem. Kategoria pliku może być jedną z trzech wartości (TEXT, IMAGE lub OTHER).
        Jeżeli stash jest tylko do odczytu, nie będzie można aktualizować kategorii
        jego zawartości.
      parameters:
      - in: path
        name: code
        required: true
        schema:
          type: string
        description: Kod stashu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                file_path:
                  type: string
                  description: Ścieżka do pliku (np. "/folder1/file.txt"). Musi zaczynać
                    się od "/" i nie może zawierać ../ lub //
                  pattern: ^(/[a-zA-Z0-9._\- ]+)+/?$
                  minLength: 1
                  maxLength: 1024
                category:
                  type: string
                  enum:
                  - TEXT
                  - IMAGE
                  - OTHER
                  description: Nowa kategoria pliku
      responses:
        '204':
          $ref: '#/components/responses/SuccessNoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

components:
  responses:
    NewStashCreated:
      description: Kod został pomyślnie utworzony
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
                description: Wygenerowany kod (pozwala wejść pod link http://localhost:8080/{code})
    StashViewSuccess:
      description: Zawartość stashu została pomyślnie zwrócona
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StashView'
    FileDownloadSuccess:
      description: Plik, folder lub archiwum zip został pomyślnie pobrany ze stashu
      content:
        application/octet-stream:
          schema:
            type: string
            format: binary
    SuccessNoContent:
      description: Operacja została pomyślnie wykonana
    BadRequest:
      description: Nieprawidłowe dane żądania lub błąd biznesowy
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Zasób nie został znaleziony
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Dostęp zabroniony - stash jest tylko do odczytu
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Konflikt - zasób o tej nazwie już istnieje
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Nieprawidłowe parametry żądania
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InsufficientStorage:
      description: Niewystarczające zasoby serwera - brak pamięci, miejsca na dysku lub inne błędy zasobów
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    StashView:
      type: object
      properties:
        valid_until:
          type: string
          format: date-time
          description: Data i czas, do którego stash jest ważny
          example: '2024-12-31T23:59:59Z'
        is_read_only:
          type: boolean
          description: Informacja, czy stash jest tylko do odczytu
        file_tree:
          $ref: '#/components/schemas/FileTreeNodeView'
    FileTreeNodeView:
      type: object
      required:
      - path
      - name
      - is_directory
      properties:
        path:
          type: string
          description: Pełna ścieżka do pliku lub katalogu
          example: /folder1/folder2/file.txt
        name:
          type: string
          description: Nazwa pliku (bez ścieżki, czyli np. "file.txt" dla ścieżki
            "/folder1/folder2/file.txt")
        is_directory:
          type: boolean
          description: Informacja, czy jest to katalog
        file:
          $ref: '#/components/schemas/FileContentView'
          description: Zawartość i metadane pliku (obecna tylko jeśli is_directory
            jest false)
        children:
          type: array
          items:
            $ref: '#/components/schemas/FileTreeNodeView'
          description: Lista dzieci węzła (obecna tylko jeśli is_directory jest true)
    FileContentView:
      type: object
      required:
      - name
      - size
      - category
      properties:
        name:
          type: string
          description: Kopia jak w FileTreeNodeView.name
        size:
          type: integer
          description: Rozmiar pliku w bajtach
        category:
          type: string
          enum:
          - TEXT
          - IMAGE
          - OTHER
          description: Kategoria pliku (TEXT, IMAGE lub OTHER)
    ErrorResponse:
      type: object
      required:
      - status
      - message
      - error
      - timestamp
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 404
        message:
          type: string
          description: Opis błędu
          example: Stash with code 'abc123' not found
        error:
          type: string
          description: Kod błędu
          example: NOT_FOUND
        timestamp:
          type: string
          format: date-time
          description: Czas zaistnienia błędu
          example: '2024-05-05T17:30:22'
        path:
          type: string
          description: Ścieżka żądania
          example: /api/v1/view/abc123
        details:
          type: string
          description: Dodatkowe szczegóły błędu (opcjonalne)
          nullable: true